lib: tool: Moved static_config_flush() to lib/static-config to make it useable for kmod.
